long

+++ main.c
@@ -6,7 +6,7 @@
 /*   By: tberube- <tberube-@student.42quebec.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/03 09:13:43 by tberube-          #+#    #+#             */
-/*   Updated: 2022/05/24 16:25:47 by tberube-         ###   ########.fr       */
+/*   Updated: 2022/05/24 16:34:05 by tberube-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -43,7 +43,7 @@ int	main(int argc, char **argv)
 	data.tab_stack = ft_calloc(2, sizeof(long *));
 	data.tab_stack[A] = ft_calloc((data.nb_arg), sizeof(long));
 	data.tab_stack[B] = ft_calloc((data.nb_arg), sizeof(long));
-	data.index = ft_calloc(data.nb_arg, sizeof(int));
+	data.index = ft_calloc(data.nb_arg, sizeof(long));
 	call_parsing(&data, argv);
 	ft_free(&data, argv);
 }
